Add files via upload
A uni project which developed an advanced drawing pad using Processing. This code provides the functions of drawing with an image as the stroke; the ability to create drawings with a variety of colours; the function to adjust the line weight; as well as save and clear the window. 

Drag mouse to draw; click mouse to change settings; using keyboard to save, clear and adjust line weight.